+++ src/components/code/code_block.test.tsx
@@ -163,6 +163,27 @@ describe('EuiCodeBlock', () => {
 
       expect(component.find('.euiCodeBlock-isFullScreen')).toMatchSnapshot();
     });
+
+    it('closes fullscreen mode when the escape key is pressed', () => {
+      const component = mount(
+        <EuiCodeBlock
+          {...requiredProps}
+          language="javascript"
+          overflowHeight={300}
+        >
+          const value = &quot;world&quot;
+        </EuiCodeBlock>
+      );
+
+      component.find('EuiButtonIcon[iconType="fullScreen"]').simulate('click');
+      component.update();
+      component
+        .find('.euiCodeBlock-isFullScreen')
+        .childAt(0)
+        .simulate('keyDown', { key: 'Escape' });
+
+      expect(component.find('.euiCodeBlock-isFullScreen')).toHaveLength(0);
+    });
   });
 
   describe('virtualization', () => {