+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Bug fixes**
 
 - Fixed `EuiCodeBlock` not passing `data-test-subj` or `aria-label` to virtualized & full-screen code blocks ([#TODO](https://github.com/elastic/eui/pull/TODO))
+- Fixed `EuiCodeBlock` not closing full-screen mode when the Escape key is pressed ([#TODO](https://github.com/elastic/eui/pull/TODO))
 
 ## [`41.1.0`](https://github.com/elastic/eui/tree/v41.1.0)